ease add/remove coders ;-)

+++ cadavre_1/test/test_next.rb
@@ -7,7 +7,7 @@ require "next.rb"
 class Next < Test::Unit::TestCase
 
   def teardown
-    File.delete("result.txt")
+    File.delete("result.txt") if File.exist?("result.txt")
   end
 
   def check_output_of_next(params, expected_exit_code, expected_result)
@@ -21,6 +21,17 @@ class Next < Test::Unit::TestCase
     
   def test_next_give_the_next_player_given_your_first_name_in_parameter
     check_output_of_next("etienne", true, "Etienne, le codeur suivant est : pascal@grange.nom.fr")
+    check_output_of_next("luc", true, "Luc, le codeur suivant est : etienne.charignon@gmail.com")
+  end
+  
+  def test_read_coder_property
+    assert_equal(["pascal@grange.nom.fr", "luc@mazardo.com", "etienne.charignon@gmail.com"], 
+      read_coders_property('coders="pascal@grange.nom.fr luc@mazardo.com etienne.charignon@gmail.com"'))
+  end
+  
+  def test_read_firstname_from_email
+    assert_equal("pascal", read_firstname("pascal@grange.nom.fr"))
+    assert_equal("etienne", read_firstname("etienne.charignon@gmail.com"))
   end
 
 end
